+++ src/Symfony/Component/PropertyAccess/Tests/PropertyPathTest.php
@@ -170,4 +170,24 @@ class PropertyPathTest extends TestCase
 
         $propertyPath->isIndex(-1);
     }
+
+    /**
+     * @dataProvider provideAppendPaths
+     */
+    public function testAppend(string $basePath, string $subPath, string $expectedPath, string $message)
+    {
+        $this->assertSame($expectedPath, PropertyPath::append($basePath, $subPath), $message);
+    }
+
+    public function provideAppendPaths()
+    {
+        return [
+            ['foo', '', 'foo', 'It returns the basePath if subPath is empty'],
+            ['', 'bar', 'bar', 'It returns the subPath if basePath is empty'],
+            ['foo', 'bar', 'foo.bar', 'It append the subPath to the basePath'],
+            ['foo', '[bar]', 'foo[bar]', 'It does not include the dot separator if subPath uses the array notation'],
+            ['0', 'bar', '0.bar', 'Leading zeros are kept.'],
+            ['0', 1, '0.1', 'Numeric subpaths do not cause PHP 7.4 errors.'],
+        ];
+    }
 }